2025/8/29 15:42 revised

+++ mindone/diffusers/pipelines/qwenimage/pipeline_qwenimage.py
@@ -196,14 +196,14 @@ class QwenImagePipeline(DiffusionPipeline, QwenImageLoraLoaderMixin):
             ""
         )
         encoder_hidden_states = self.text_encoder(
-            input_ids=ms.Tensor(txt_tokens.input_ids, dtype=ms.int32),
-            attention_mask=ms.Tensor(txt_tokens.attention_mask, dtype=ms.int32),
+            input_ids=ms.Tensor(txt_tokens.input_ids),
+            attention_mask=ms.Tensor(txt_tokens.attention_mask),
             output_hidden_states=True,
         )
         hidden_states = encoder_hidden_states.hidden_states[-1]
         split_hidden_states = self._extract_masked_hidden(hidden_states, ms.Tensor(txt_tokens.attention_mask))
         split_hidden_states = [e[drop_idx:] for e in split_hidden_states]
-        attn_mask_list = [mint.ones(e.shape[0], dtype=ms.int32) for e in split_hidden_states]
+        attn_mask_list = [mint.ones(e.shape[0], dtype=ms.int64) for e in split_hidden_states]
         max_seq_len = max([e.shape[0] for e in split_hidden_states])
         prompt_embeds = mint.stack(
             [mint.cat([u, u.new_zeros((max_seq_len - u.shape[0], u.shape[1]))]) for u in split_hidden_states]
